quiz_backend/requirements.txt
fastapi>=0.115.8
greenlet==2.0.2
haystack-ai==2.2.0
json-repair>=0.39.1
pydantic>=2.10.6
pypdf>=5.3.1
python-dotenv>=1.0.1
python-multipart>=0.0.20
requests>=2.32.3
sqlalchemy>=2.0.38
uvicorn[standard]>=0.34.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
stripe>=7.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 